refactor: 프로젝트 분리(step1)
