+++ src/main/kotlin/study/Person.kt
@@ -1,0 +1,9 @@
+package study
+
+class Person(
+    val name: String,
+    val company: String? = null,
+    val skills: Skills = Skills(emptyList(), emptyList()),
+    val languages: List<Language> = emptyList()
+
+)
