+++ src/main/kotlin/study/Skills.kt
@@ -1,0 +1,3 @@
+package study
+
+data class Skills(val softSkills: List<String>, val hardSkills: List<String>)
