+++ src/main/kotlin/study/SkillsBuilder.kt
@@ -1,0 +1,18 @@
+package study
+
+class SkillsBuilder {
+    var softSkills: MutableList<String> = mutableListOf()
+    var hardSkills: MutableList<String> = mutableListOf()
+
+    fun soft(value: String) {
+        softSkills.add(value)
+    }
+
+    fun hard(value: String) {
+        hardSkills.add(value)
+    }
+
+    fun build(): Skills {
+        return Skills(softSkills, hardSkills)
+    }
+}
